+++ frontend/app/(main)/practice/commands.ts
@@ -1,2 +1,23 @@
-//backend API has a submit review route already that takes the tune ID and the feedback quality as a string. 
-//Doesn't currently allow for updating the notes, I don't think. 
+import axios from 'axios';
+
+const client = axios.create({
+    baseURL: "http://127.0.0.1:8000/tunetrees" 
+  });
+
+interface PracticeFeedbackProps {
+    id: string
+    feedback: string
+}
+
+export const submitPracticeFeedback = async ({id, feedback}: PracticeFeedbackProps) => {
+    try {
+        const stringified = JSON.stringify({
+            selected_tune: id, vote_type: feedback
+        })
+
+        await client.post("/practice/feedback", {body: stringified})
+    }
+    catch(e: any) {
+        console.log("Unable to post feedback.")
+    }
+}
